move subscription to redux
